functions/generate_pdfs/requirements.txt
functions-framework==3.0.0
google-cloud-bigquery[bqstorage,pandas]==3.7.0
google-cloud-storage==2.7.0
pandas==1.5.3
weasyprint==52.5